+++ prover/benches/bench_kzg_commit.rs
@@ -1,18 +1,18 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use rand::Rng;
-use rust_kzg_bn254::{blob::Blob, kzg::KZG, srs::SRS};
+use rust_kzg_bn254_primitives::blob::Blob;
+use rust_kzg_bn254_prover::{kzg::KZG, srs::SRS};
 use std::time::Duration;
 
 fn bench_kzg_commit(c: &mut Criterion) {
     let mut rng = rand::thread_rng();
-    let mut kzg = KZG::new(
-        SRS::new(
-            "tests/test-files/mainnet-data/g1.131072.point",
-            268435456,
-            131072,
-        )
-        .unwrap(),
-    );
+    let mut kzg = KZG::new();
+    let srs = SRS::new(
+        "tests/test-files/mainnet-data/g1.32mb.point",
+        268435456,
+        524288,
+    )
+    .unwrap();
 
     c.bench_function("bench_kzg_commit_10000", |b| {
         let random_blob: Vec<u8> = (0..10000).map(|_| rng.gen_range(32..=126) as u8).collect();
@@ -20,7 +20,7 @@ fn bench_kzg_commit(c: &mut Criterion) {
         let input_poly = input.to_polynomial_coeff_form();
         kzg.calculate_and_store_roots_of_unity(input.len().try_into().unwrap())
             .unwrap();
-        b.iter(|| kzg.commit_coeff_form(&input_poly).unwrap());
+        b.iter(|| kzg.commit_coeff_form(&input_poly, &srs).unwrap());
     });
 
     c.bench_function("bench_kzg_commit_30000", |b| {
@@ -29,7 +29,7 @@ fn bench_kzg_commit(c: &mut Criterion) {
         let input_poly = input.to_polynomial_coeff_form();
         kzg.calculate_and_store_roots_of_unity(input.len().try_into().unwrap())
             .unwrap();
-        b.iter(|| kzg.commit_coeff_form(&input_poly).unwrap());
+        b.iter(|| kzg.commit_coeff_form(&input_poly, &srs).unwrap());
     });
 
     c.bench_function("bench_kzg_commit_50000", |b| {
@@ -38,7 +38,7 @@ fn bench_kzg_commit(c: &mut Criterion) {
         let input_poly = input.to_polynomial_coeff_form();
         kzg.calculate_and_store_roots_of_unity(input.len().try_into().unwrap())
             .unwrap();
-        b.iter(|| kzg.commit_coeff_form(&input_poly).unwrap());
+        b.iter(|| kzg.commit_coeff_form(&input_poly, &srs).unwrap());
     });
 }
