+++ prover/benches/bench_kzg_proof.rs
@@ -1,18 +1,18 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use rand::Rng;
-use rust_kzg_bn254::{blob::Blob, kzg::KZG, srs::SRS};
+use rust_kzg_bn254_primitives::blob::Blob;
+use rust_kzg_bn254_prover::{kzg::KZG, srs::SRS};
 use std::time::Duration;
 
 fn bench_kzg_proof(c: &mut Criterion) {
     let mut rng = rand::thread_rng();
-    let mut kzg = KZG::new(
-        SRS::new(
-            "tests/test-files/mainnet-data/g1.131072.point",
-            268435456,
-            131072,
-        )
-        .unwrap(),
-    );
+    let mut kzg = KZG::new();
+    let srs = SRS::new(
+        "tests/test-files/mainnet-data/g1.32mb.point",
+        268435456,
+        524288,
+    )
+    .unwrap();
 
     c.bench_function("bench_kzg_proof_10000", |b| {
         let random_blob: Vec<u8> = (0..10000).map(|_| rng.gen_range(32..=126) as u8).collect();
@@ -23,7 +23,7 @@ fn bench_kzg_proof(c: &mut Criterion) {
         let index =
             rand::thread_rng().gen_range(0..input_poly.len_underlying_blob_field_elements());
         b.iter(|| {
-            kzg.compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap())
+            kzg.compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap(), &srs)
                 .unwrap()
         });
     });
@@ -37,7 +37,7 @@ fn bench_kzg_proof(c: &mut Criterion) {
         let index =
             rand::thread_rng().gen_range(0..input_poly.len_underlying_blob_field_elements());
         b.iter(|| {
-            kzg.compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap())
+            kzg.compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap(), &srs)
                 .unwrap()
         });
     });
@@ -51,7 +51,7 @@ fn bench_kzg_proof(c: &mut Criterion) {
         let index =
             rand::thread_rng().gen_range(0..input_poly.len_underlying_blob_field_elements());
         b.iter(|| {
-            kzg.compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap())
+            kzg.compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap(), &srs)
                 .unwrap()
         });
     });
