+++ prover/kzg_commitment_diagram.png
