+++ prover/tests/test-files/mainnet-data/g2.point.powerOf2
