+++ verifier/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rust-kzg-bn254-batch-verification"
+name = "rust-kzg-bn254-batch-verifier"
 version = "0.1.0"
 license-file.workspace = true
 edition.workspace = true
@@ -20,4 +20,9 @@ lazy_static = "1.5"
 tracing = { version = "0.1.41", features = ["log"] }
 tracing-subscriber = "0.3.19"
 ark-std = { version = "0.5.0", features = ["parallel"] }
-rust-kzg-bn254 = { path = "../kzg", version = "0.1.0" }
+rust-kzg-bn254-prover = { path = "../prover", version = "0.1.0" }
+
+[[bench]]
+name = "bench_kzg_verify"
+harness = false
+path = "benches/bench_kzg_verify.rs"
