+++ verifier/benches/bench_kzg_verify.rs
@@ -1,18 +1,19 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use rand::Rng;
-use rust_kzg_bn254::{blob::Blob, kzg::KZG, srs::SRS, verify};
+use rust_kzg_bn254_batch_verifier::verify::verify_proof;
+use rust_kzg_bn254_primitives::blob::Blob;
+use rust_kzg_bn254_prover::{kzg::KZG, srs::SRS};
 use std::time::Duration;
 
 fn bench_kzg_verify(c: &mut Criterion) {
     let mut rng = rand::thread_rng();
-    let mut kzg = KZG::new(
-        SRS::new(
-            "tests/test-files/mainnet-data/g1.131072.point",
-            268435456,
-            131072,
-        )
-        .unwrap(),
-    );
+    let mut kzg = KZG::new();
+    let srs = SRS::new(
+        "../prover/tests/test-files/mainnet-data/g1.32mb.point",
+        268435456,
+        524288,
+    )
+    .unwrap();
 
     c.bench_function("bench_kzg_verify_10000", |b| {
         let random_blob: Vec<u8> = (0..10000).map(|_| rng.gen_range(32..=126) as u8).collect();
@@ -22,13 +23,13 @@ fn bench_kzg_verify(c: &mut Criterion) {
             .unwrap();
         let index =
             rand::thread_rng().gen_range(0..input_poly.len_underlying_blob_field_elements());
-        let commitment = kzg.commit_eval_form(&input_poly).unwrap();
+        let commitment = kzg.commit_eval_form(&input_poly, &srs).unwrap();
         let proof = kzg
-            .compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap())
+            .compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap(), &srs)
             .unwrap();
         let value_fr = input_poly.get_evalualtion(index).unwrap();
         let z_fr = kzg.get_nth_root_of_unity(index).unwrap();
-        b.iter(|| verify::proof(commitment, proof, *value_fr, *z_fr));
+        b.iter(|| verify_proof(commitment, proof, *value_fr, *z_fr));
     });
 
     c.bench_function("bench_kzg_verify_30000", |b| {
@@ -39,13 +40,13 @@ fn bench_kzg_verify(c: &mut Criterion) {
             .unwrap();
         let index =
             rand::thread_rng().gen_range(0..input_poly.len_underlying_blob_field_elements());
-        let commitment = kzg.commit_eval_form(&input_poly).unwrap();
+        let commitment = kzg.commit_eval_form(&input_poly, &srs).unwrap();
         let proof = kzg
-            .compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap())
+            .compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap(), &srs)
             .unwrap();
         let value_fr = input_poly.get_evalualtion(index).unwrap();
         let z_fr = kzg.get_nth_root_of_unity(index).unwrap();
-        b.iter(|| verify::proof(commitment, proof, *value_fr, *z_fr));
+        b.iter(|| verify_proof(commitment, proof, *value_fr, *z_fr));
     });
 
     c.bench_function("bench_kzg_verify_50000", |b| {
@@ -56,13 +57,13 @@ fn bench_kzg_verify(c: &mut Criterion) {
             .unwrap();
         let index =
             rand::thread_rng().gen_range(0..input_poly.len_underlying_blob_field_elements());
-        let commitment = kzg.commit_eval_form(&input_poly).unwrap();
+        let commitment = kzg.commit_eval_form(&input_poly, &srs).unwrap();
         let proof = kzg
-            .compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap())
+            .compute_proof_with_known_z_fr_index(&input_poly, index.try_into().unwrap(), &srs)
             .unwrap();
         let value_fr = input_poly.get_evalualtion(index).unwrap();
         let z_fr = kzg.get_nth_root_of_unity(index).unwrap();
-        b.iter(|| verify::proof(commitment, proof, *value_fr, *z_fr));
+        b.iter(|| verify_proof(commitment, proof, *value_fr, *z_fr));
     });
 }
